# Optimize IDNA tables